+++ bench/bench_axpyv.c
@@ -236,7 +236,7 @@ int main( int argc, char** argv )
 
 			dtime_save = bli_clock_min_diff( dtime_save, dtime );
 		}
-		gflops = ( 3.0 * n ) / ( dtime_save * 1.0e9 );
+		gflops = ( 3.0 * (double)n ) / ( dtime_save * 1.0e9 );
 		if ( bli_is_complex( dt ) ) gflops *= 4.0;
 
 		printf( "data_axpyv_%s", BLAS );
